Move command to another VM
