+++ src/DevSpector.UI/ViewModels/DevicesSection/Implementations/FreeIPListViewModel.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using DevSpector.Desktop.Service;
 using ReactiveUI;
+using DevSpector.SDK.Models;
 
 namespace DevSpector.Desktop.UI.ViewModels
 {
@@ -25,7 +26,23 @@ namespace DevSpector.Desktop.UI.ViewModels
             _storage = storage;
             _messagesBroker = messagesBroker;
             _devicesListViewModel = devicesListViewModel;
+
+            AddFreeIPToDeviceCommand = ReactiveCommand.CreateFromTask(
+                AddIPToDeviceAsync,
+                this.WhenAny(
+                    (vm) => vm.SelectedItem,
+                    (vm) => vm._devicesListViewModel.SelectedItem,
+                    (freeIp, selectedDevice) => {
+                        if (SelectedItem == null) return false;
+                        if (_devicesListViewModel.SelectedItem == null) return false;
+                        return true;
+                    }
+                )
+            );
+
         }
+
+        public ReactiveCommand<Unit, Unit> AddFreeIPToDeviceCommand { get; }
 
         public override string SelectedItem
         {
@@ -54,5 +71,24 @@ namespace DevSpector.Desktop.UI.ViewModels
                 _messagesBroker.NotifyUser(e.Message);
             }
         }
+
+        public async Task AddIPToDeviceAsync()
+        {
+            try
+            {
+                Device selectedDevice = _devicesListViewModel.SelectedItem;
+
+                await _storage.AddIPAsync(selectedDevice.InventoryNumber, SelectedItem);
+
+                _messagesBroker.NotifyUser(
+                    $"IP-адрес \"{SelectedItem}\" был добавлен к устройству \"{selectedDevice.InventoryNumber}\""
+                );
+            }
+            catch (Exception e)
+            {
+                _messagesBroker.NotifyUser(e.Message);
+            }
+        }
+
     }
 }
